handleMatch now can run without re-rendering the GameBoard
